Fix failing tests from previous commit
