+++ SIL.DictionaryServices.Tests/Model/LexRelationTests.cs
@@ -43,35 +43,6 @@ namespace SIL.DictionaryServices.Tests.Model
 	}
 
 	[TestFixture]
-	public class LexRelationCollectionCloneableTests : CloneableTests<IPalasoDataObjectProperty>
-	{
-		public override IPalasoDataObjectProperty CreateNewCloneable()
-		{
-			return new LexRelationCollection();
-		}
-
-		public override string ExceptionList
-		{
-			//PropertyChanged: No good way to clone eventhandlers
-			//_parent: We are doing top down clones. Children shouldn't make clones of their parents, but parents of their children.
-			get { return "|_parent|PropertyChanged|"; }
-		}
-
-		protected override List<ValuesToSet> DefaultValuesForTypes
-		{
-			get
-			{
-				return new List<ValuesToSet>
-							 {
-								 new ValuesToSet(
-									new List<LexRelation> { new LexRelation("id", "target", null) },
-									new List<LexRelation> { new LexRelation("!id", "!target", null) }),
-							 };
-			}
-		}
-	}
-
-	[TestFixture]
 	public class LexRelationTests
 	{
 		private string _filePath;
