+++ SIL.DictionaryServices.Tests/Model/LexTraitTests.cs
@@ -1,9 +1,7 @@
 ﻿using System.Collections.Generic;
 using NUnit.Framework;
 using SIL.DictionaryServices.Model;
-using SIL.Lift;
 using SIL.TestUtilities;
-using SIL.Text;
 
 namespace SIL.DictionaryServices.Tests.Model
 {
@@ -15,53 +13,12 @@ namespace SIL.DictionaryServices.Tests.Model
 			return new LexTrait("type", "value");
 		}
 
-		public override string ExceptionList
-		{
-			get { return ""; }
-		}
+		public override string ExceptionList => "";
 
-		protected override List<ValuesToSet> DefaultValuesForTypes
-		{
-			get
+		protected override List<ValuesToSet> DefaultValuesForTypes =>
+			new List<ValuesToSet>
 			{
-				return new List<ValuesToSet>
-							 {
-								 new ValuesToSet("to be", "!(to be)")
-							 };
-			}
-		}
-	}
-
-	[TestFixture]
-	public class LexFieldCloneableTests : CloneableTests<IPalasoDataObjectProperty>
-	{
-		public override IPalasoDataObjectProperty CreateNewCloneable()
-		{
-			return new LexField("type");
-		}
-
-		public override string ExceptionList
-		{
-			//PropertyChanged: No good way to clone eventhandlers
-			//_parent: We are doing top down clones. Children shouldn't make clones of their parents, but parents of their children.
-			get { return "|_parent|PropertyChanged|"; }
-		}
-
-		protected override List<ValuesToSet> DefaultValuesForTypes
-		{
-			get
-			{
-				return new List<ValuesToSet>
-						   {
-							   new ValuesToSet("to be", "!(to be)"),
-							   new ValuesToSet(new[] {new LanguageForm("en", "en_form", null)},
-												 new[] {new LanguageForm("de", "de_form", null)}),
-							   new ValuesToSet(
-								   new List<LexTrait> {new LexTrait("one", "eins"), new LexTrait("two", "zwei")},
-								   new List<LexTrait> {new LexTrait("three", "drei"), new LexTrait("four", "vier")}),
-							   new ValuesToSet(new List<string> {"to", "be"}, new List<string> {"!", "to", "be"})
-						   };
-			}
-		}
+				new ValuesToSet("to be", "!(to be)")
+			};
 	}
 }
